% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acv_arma.R
\name{acv_arma}
\alias{acv_arma}
\title{Compute a theoretical autocovariance function of ARMA process}
\usage{
acv_arma(phi, theta, n)
}
\arguments{
\item{phi}{vector containing the AR parameters}

\item{theta}{vector containing the MA parameters}

\item{n}{length of the time series}
}
\value{
vector of length n containing the autocovariances
}
\description{
Function \code{acv_arma} computes a theoretical autocovariance function of ARMA process.
}
\seealso{
\code{\link{dacv_arma}}.
}

