% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arima_pi.R
\name{arima_pi}
\alias{arima_pi}
\title{Compute Bayesian prediction intervals for ARIMA processes with exogenous variables using importance sampling}
\usage{
arima_pi(y, order, xreg = NULL, n.ahead = 1, level = 0.95,
  median = TRUE, se_limits = TRUE, priors = c("uniform",
  "approx_joint_jeffreys", "approx_marginal_jeffreys", "exact_joint_jeffreys",
  "exact_marginal_jeffreys"), custom_prior, custom_prior_args, nsim = 1000,
  last_only = FALSE, return_weights = FALSE, ...)
}
\arguments{
\item{y}{vector containing the time series}

\item{order}{vector of length 3 with values p,d,q
corresponding to the number of AR parameters, degree of differencing and number of MA parameters.}

\item{xreg}{matrix or data frame containing the exogenous variables
(not including the intercept which is always included for non-differenced series)}

\item{n.ahead}{length of the forecast horizon.}

\item{level}{desired frequentist coverage probability of the prediction intervals.}

\item{median}{compute the median of the prediction interval.}

\item{se_limits}{compute the standard errors of the prediction interval limits.}

\item{priors}{priors to be used in importance sampling. Multiple choices are allowed. See \code{\link{jeffreys}} for details.}

\item{custom_prior}{function for computing custom prior.
First argument must be a vector containing the AR and MA parameters (in that order).}

\item{custom_prior_args}{list containing additional arguments to \code{custom_prior}.}

\item{nsim}{number of simulations used in importance sampling.}

\item{last_only}{compute the prediction intervals only for the last prediction step.}

\item{return_weights}{Return (scaled) weights used in importance sampling.}

\item{...}{Additional arguments for \code{arima}.}
}
\value{
a list containing the prediction intervals.
}
\description{
Function \code{arima_pi} computes prediction intervals for ARIMA processes
with exogenous variables using importance sampling. For regression coefficients,
diffuse (uninformative) prior is used, whereas multiple options for
prior distributions for ARMA coefficients are supported.
}
\examples{
pred_arima <- predict(arima(lh, order = c(3,0,0)), n.ahead = 12, se.fit = TRUE)
pred_arima <- cbind(pred = pred_arima$pred,
 lwr = pred_arima$pred - qnorm(0.975)*pred_arima$se^2,
 upr = pred_arima$pred + qnorm(0.975)*pred_arima$se^2)

pred <- arima_pi(lh, order = c(3,0,0), n.ahead = 12, prior = "uniform")

ts.plot(ts.union(lh,pred_arima, pred$uniform[,1:3]), col = c(1,2,2,2,3,3,3),
lty = c(1,1,2,2,1,2,2))
}
\references{
\enumerate{
\item{Helske, J. and Nyblom, J. (2013). Improved frequentist prediction
intervals for autoregressive models by simulation.
In Siem Jan Koopman and Neil Shephard, editors,
Unobserved Components and Time Series Econometrics. Oxford University Press. In press.}
 \item{Helske, J. and Nyblom, J. (2014). Improved frequentist prediction intervals for
 ARMA models by simulation.
 In Johan Knif and Bernd Pape, editors,
Contributions to Mathematics, Statistics, Econometrics, and Finance:
essays in honour of professor Seppo Pynnonen,
number 296 in Acta Wasaensia, pages 71–86. University of Vaasa.}
}
}

