% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acv_arma.R
\name{dacv_arma}
\alias{dacv_arma}
\title{Compute the partial derivatives of theoretical autocovariance function of ARMA process}
\usage{
dacv_arma(phi, theta, n)
}
\arguments{
\item{phi}{vector containing the AR parameters}

\item{theta}{vector containing the MA parameters}

\item{n}{length of the time series}
}
\value{
matrix containing the partial derivatives autocovariances,
each column corresponding to one parameter of vector (phi,theta) (in that order)
}
\description{
Function \code{dacv_arma} computes the partial derivatives of theoretical autocovariance function of ARMA process
}
\seealso{
\code{\link{acv_arma}}.
}

