% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_arma.R
\name{information_arma}
\alias{information_arma}
\title{Large Sample Approximation of Information Matrix for ARMA process}
\usage{
information_arma(phi = NULL, theta = NULL)
}
\arguments{
\item{phi}{Autoregressive coefficients.}

\item{theta}{Moving average coefficients.}
}
\value{
Large sample approximation of information matrix for ARMA process.
}
\description{
Fortran implementation of \code{InformationMatrixARMA} function of
\code{FitARMA} package, except that the function uses the same
ARMA model definition as \code{arima}, where both the
AR and MA parts of the model are on the right side of the equation, i.e.
MA coefficients differ in sign compared to \code{InformationMatrixARMA}.
}
\references{
\enumerate{
\item{Box, G. and Jenkins, G. (1970). Time Series Analysis: Forecasting and Control. San Francisco: Holden-Day.}
\item{McLeod, A. I. and Zhang, Y., (2007). Faster ARMA maximum likelihood estimation Computational Statistics & Data
Analysis 52(4) URL http://dx.doi.org/10.1016/j.csda.2007.07.020}
}
}

