% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avg_coverage_struct.R
\name{avg_coverage_struct}
\alias{avg_coverage_struct}
\title{Compute the average coverage of the prediction intervals computed by
naive plug-in method and \code{\link{struct_pi}}}
\usage{
avg_coverage_struct(type = c("level", "trend", "BSM"), sds, n, n_ahead = 1,
  nsim2, nsim = 100, level = 0.95, prior = "uniform", custom_prior,
  custom_prior_args, return_all_coverages = FALSE, ...)
}
\arguments{
\item{type}{Type of model. See \code{\link{struct_pi}}.}

\item{sds}{vector containing the standard deviations of the model.}

\item{n}{length of the time series}

\item{n_ahead}{length of the forecast horizon}

\item{nsim2}{number of simulations used in computing the expected coverage}

\item{nsim}{number of simulations used in importance sampling}

\item{level}{desired coverage probability of the prediction intervals}

\item{prior}{prior to be used in importance sampling.}

\item{custom_prior}{function for computing custom prior. First argument must be a vector containing the AR and MA parameters (in that order).}

\item{custom_prior_args}{list containing additional arguments to \code{custom_prior}.}

\item{return_all_coverages}{return raw results i.e. coverages for each simulations. When \code{FALSE} (default), summary statistics are returned.}

\item{...}{Optional arguments to \code{\link{arima_pi}}.}
}
\value{
a list containing the coverage probabilities
}
\description{
Computes expected coverage probabilities of the prediction intervals of structural time series model.
}
\seealso{
\code{\link{struct_pi}}.
}

